# backend/openapi/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/amenities": {
      "get": {
        "operationId": "AmenitiesController_list",
        "parameters": [
          {
            "name": "category",
            "required": false,
            "in": "query",
            "description": "Filter by category",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Search by name (ILIKE)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number (1-based)",
            "schema": {
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "description": "Page size",
            "schema": {
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "description": "Sort field",
            "schema": {
              "default": "name",
              "type": "string",
              "enum": ["name", "createdAt"]
            }
          },
          {
            "name": "sortOrder",
            "required": false,
            "in": "query",
            "description": "Sort order",
            "schema": {
              "default": "asc",
              "type": "string",
              "enum": ["asc", "desc"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of amenities"
          }
        },
        "summary": "List amenities",
        "tags": ["amenities"]
      },
      "post": {
        "operationId": "AmenitiesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAmenityDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Amenity created"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create amenity",
        "tags": ["amenities"]
      }
    },
    "/v1/amenities/{id}": {
      "patch": {
        "operationId": "AmenitiesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAmenityDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Amenity updated"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update amenity",
        "tags": ["amenities"]
      },
      "delete": {
        "operationId": "AmenitiesController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Amenity deleted"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete amenity",
        "tags": ["amenities"]
      }
    },
...
    "/v1/hotel-packages": {
      "get": {
        "operationId": "HotelPackagesController_list",
        "parameters": [
          {
            "name": "providerId",
            "required": false,
            "in": "query",
            "description": "Filter by provider ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "Filter by status",
            "schema": {
              "type": "string",
              "enum": ["DRAFT", "PUBLISHED", "PAUSED", "ARCHIVED"]
            }
          },
          {
            "name": "templateId",
            "required": false,
            "in": "query",
            "description": "Filter by template ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "listingId",
            "required": false,
            "in": "query",
            "description": "Filter by listing ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Search by name (ILIKE)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "includeArchived",
            "required": false,
            "in": "query",
            "description": "Include archived packages",
            "schema": {
              "default": false,
              "type": "boolean"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number (1-based)",
            "schema": {
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "description": "Page size",
            "schema": {
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "description": "Sort field",
            "schema": {
              "default": "createdAt",
              "type": "string",
              "enum": ["createdAt", "name"]
            }
          },
          {
            "name": "sortOrder",
            "required": false,
            "in": "query",
            "description": "Sort order",
            "schema": {
              "default": "desc",
              "type": "string",
              "enum": ["asc", "desc"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of hotel packages"
          }
        },
        "summary": "List hotel packages",
        "tags": ["hotel-packages"]
      }
    },
...
    "/v1/hotel-packages/{id}": {
      "get": {
        "operationId": "HotelPackagesController_getOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Hotel package detail with amenities"
          }
        },
        "summary": "Get hotel package by id",
        "tags": ["hotel-packages"]
      }
    },
    "/v1/hotel-packages/{providerId}/packages": {
      "post": {
        "operationId": "HotelPackagesController_createHotelPackage",
        "parameters": [
          {
            "name": "providerId",
            "required": true,
            "in": "path",
            "description": "ProviderProfile id",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateHotelPackageDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Hotel package created in DRAFT status"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create hotel package",
        "tags": ["hotel-packages"]
      }
    },
    "/v1/hotel-packages/{providerId}/packages/{packageId}/publish": {
      "post": {
        "operationId": "HotelPackagesController_publish",
        "parameters": [
          {
            "name": "providerId",
            "required": true,
            "in": "path",
            "description": "ProviderProfile id",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "packageId",
            "required": true,
            "in": "path",
            "description": "HotelPackage id",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Hotel package published"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Publish a hotel package (requires verified provider)",
        "tags": ["hotel-packages"]
      }
    },
    "/v1/hotel-packages/{providerId}/packages/{packageId}/pause": {
      "post": {
        "operationId": "HotelPackagesController_pause",
        "parameters": [
          {
            "name": "providerId",
            "required": true,
            "in": "path",
            "description": "ProviderProfile id",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "packageId",
            "required": true,
            "in": "path",
            "description": "HotelPackage id",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Hotel package paused"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Pause a published hotel package",
        "tags": ["hotel-packages"]
      }
    },
    "/v1/hotel-packages/{providerId}/packages/{packageId}/archive": {
      "post": {
        "operationId": "HotelPackagesController_archive",
        "parameters": [
          {
            "name": "providerId",
            "required": true,
            "in": "path",
            "description": "ProviderProfile id",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "packageId",
            "required": true,
            "in": "path",
            "description": "HotelPackage id",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Hotel package archived"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Archive a hotel package",
        "tags": ["hotel-packages"]
      }
    },
    "/v1/hotel-packages/{providerId}/packages/{packageId}": {
      "patch": {
        "operationId": "HotelPackagesController_update",
        "parameters": [
          {
            "name": "providerId",
            "required": true,
            "in": "path",
            "description": "ProviderProfile id",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "packageId",
            "required": true,
            "in": "path",
            "description": "HotelPackage id",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateHotelPackageDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Hotel package updated"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update a hotel package (DRAFT only)",
        "tags": ["hotel-packages"]
      },
      "delete": {
        "operationId": "HotelPackagesController_remove",
        "parameters": [
          {
            "name": "providerId",
            "required": true,
            "in": "path",
            "description": "ProviderProfile id",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "packageId",
            "required": true,
            "in": "path",
            "description": "HotelPackage id",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Hotel package deleted"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete a hotel package (DRAFT only)",
        "tags": ["hotel-packages"]
      }
    },
...
  },
  "info": {
    "title": "TripAvail API",
    "description": "Two-sided travel marketplace API",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:4100",
      "description": "Development"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateAmenityDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Free WiFi"
          },
          "category": {
            "type": "string",
            "example": "Room"
          }
        },
        "required": ["name"]
      },
      "UpdateAmenityDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Free WiFi"
          },
          "category": {
            "type": "string",
            "example": "Room"
          }
        }
      },
      "CreateHotelPackageDto": {
        "type": "object",
        "properties": {
          "templateId": {
            "type": "string",
            "example": "weekend-getaway",
            "enum": [
              "weekend-getaway",
              "honeymoon",
              "family-vacation",
              "corporate-retreat",
              "ski-snow-holiday",
              "golf-getaway",
              "staycation",
              "wellness-spa",
              "cultural-experience",
              "food-wine",
              "anniversary",
              "eco-nature",
              "festival-events",
              "adventure"
            ]
          },
          "listingId": {
            "type": "string",
            "example": "listing_cuid"
          },
          "name": {
            "type": "string",
            "example": "Romantic Weekend"
          },
          "description": {
            "type": "string",
            "example": "Two nights with breakfast and spa access."
          },
          "pricePerPerson": {
            "type": "number",
            "example": 399.99
          },
          "inclusions": {
            "example": ["Breakfast", "Spa access"],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "exclusions": {
            "example": ["Flights not included"],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "availabilityRule": {
            "type": "string",
            "example": "WEEKEND_ONLY",
            "enum": ["WEEKEND_ONLY", "SEASONAL", "FLEXIBLE"]
          },
          "amenityIds": {
            "example": ["amenity_cuid1", "amenity_cuid2"],
            "description": "Amenity IDs to attach",
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "templateId",
          "listingId",
          "name",
          "description",
          "pricePerPerson",
          "availabilityRule"
        ]
      },
      "UpdateHotelPackageDto": {
        "type": "object",
        "properties": {
          "templateId": {
            "type": "string",
            "example": "weekend-getaway",
            "enum": [
              "weekend-getaway",
              "honeymoon",
              "family-vacation",
              "corporate-retreat",
              "ski-snow-holiday",
              "golf-getaway",
              "staycation",
              "wellness-spa",
              "cultural-experience",
              "food-wine",
              "anniversary",
              "eco-nature",
              "festival-events",
              "adventure"
            ]
          },
          "name": {
            "type": "string",
            "example": "Romantic Weekend"
          },
          "description": {
            "type": "string",
            "example": "Two nights with breakfast and spa access."
          },
          "pricePerPerson": {
            "type": "number",
            "example": 399.99
          },
          "inclusions": {
            "example": ["Breakfast", "Spa access"],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "exclusions": {
            "example": ["Flights not included"],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "availabilityRule": {
            "type": "string",
            "example": "WEEKEND_ONLY",
            "enum": ["WEEKEND_ONLY", "SEASONAL", "FLEXIBLE"]
          },
          "amenityIds": {
            "example": ["amenity_cuid1", "amenity_cuid2"],
            "description": "Replace amenities on the package (DRAFT only)",
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
    }
  }
}